split remote to common protocol package and engine implemetation
